Repackaging
- moved package 'org.cufy.http.syntax' to 'org.cufy.internal.syntax'
- moved uri components to package 'org.cufy.uri'
