+++ src/main/java/org/cufy/http/Method.java
@@ -15,8 +15,8 @@
  */
 package org.cufy.http;
 
-import org.cufy.http.syntax.HttpPattern;
-import org.cufy.http.syntax.HttpRegExp;
+import org.cufy.internal.syntax.HttpPattern;
+import org.cufy.internal.syntax.HttpRegExp;
 import org.intellij.lang.annotations.Pattern;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
