+++ src/main/java/org/cufy/http/Port.java
@@ -15,8 +15,8 @@
  */
 package org.cufy.http;
 
-import org.cufy.http.syntax.UriPattern;
-import org.cufy.http.syntax.UriRegExp;
+import org.cufy.internal.syntax.UriPattern;
+import org.cufy.internal.syntax.UriRegExp;
 import org.intellij.lang.annotations.Pattern;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
