+++ src/main/java/org/cufy/http/Response.java
@@ -16,10 +16,10 @@
 package org.cufy.http;
 
 import org.cufy.http.body.BytesBody;
-import org.cufy.http.syntax.AbnfPattern;
-import org.cufy.http.syntax.HttpParse;
-import org.cufy.http.syntax.HttpPattern;
-import org.cufy.http.syntax.HttpRegExp;
+import org.cufy.internal.syntax.AbnfPattern;
+import org.cufy.internal.syntax.HttpParse;
+import org.cufy.internal.syntax.HttpPattern;
+import org.cufy.internal.syntax.HttpRegExp;
 import org.intellij.lang.annotations.Pattern;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
