+++ src/main/java/org/cufy/uri/Path.java
@@ -13,10 +13,10 @@
  *	See the License for the specific language governing permissions and
  *	limitations under the License.
  */
-package org.cufy.http;
+package org.cufy.uri;
 
-import org.cufy.http.syntax.UriPattern;
-import org.cufy.http.syntax.UriRegExp;
+import org.cufy.internal.syntax.UriPattern;
+import org.cufy.internal.syntax.UriRegExp;
 import org.intellij.lang.annotations.Pattern;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
@@ -28,77 +28,73 @@ import java.util.Objects;
 /**
  * <b>Constant</b> (PCT Encode)
  * <br>
- * The "Host" part of the "Authority" part of a Uri.
+ * The "Path" part of a Uri.
  *
  * @author LSafer
  * @version 0.0.1
  * @since 0.0.1 ~2021.03.20
  */
-public class Host implements Serializable {
+public class Path implements Serializable {
 	/**
-	 * Local host constant.
+	 * Unspecified path constant.
+	 *
+	 * @since 0.1.0 ~2021.08.17
+	 */
+	public static final Path UNSPECIFIED = new Path("");
+
+	@SuppressWarnings("JavaDoc")
+	private static final long serialVersionUID = -5564283348900202148L;
+
+	/**
+	 * The path literal.
 	 *
 	 * @since 0.0.1 ~2021.03.21
 	 */
-	public static final Host LOCALHOST = new Host("localhost");
-	/**
-	 * Unspecified host constant.
-	 *
-	 * @since 0.3.0 ~2021.11.16
-	 */
-	public static final Host UNSPECIFIED = new Host("");
-
-	@SuppressWarnings("JavaDoc")
-	private static final long serialVersionUID = 3437183298806216312L;
-
-	/**
-	 * The string value of this host.
-	 *
-	 * @since 0.0.1 ~2021.03.20
-	 */
 	@NotNull
-	@Pattern(UriRegExp.HOST)
+	@Pattern(UriRegExp.PATH)
 	protected final String value;
 
 	/**
-	 * Construct a new default-implementation host from the given {@code source}.
+	 * Construct a new default-implementation path component with its path literal being
+	 * the given {@code source}.
 	 * <br>
 	 * Note: No validation will be applied.
 	 *
-	 * @param source the source of the constructed host.
+	 * @param source the source of the path literal of the constructed path component.
 	 * @throws NullPointerException if the given {@code source} is null.
-	 * @since 0.0.1 ~2021.03.20
+	 * @since 0.0.1 ~2021.03.21
 	 */
-	public Host(@NotNull @Pattern(UriRegExp.HOST) String source) {
+	public Path(@NotNull @Pattern(UriRegExp.PATH) String source) {
 		Objects.requireNonNull(source, "source");
 		this.value = source;
 	}
 
 	/**
-	 * Create a new host from parsing the given {@code source}.
+	 * Construct a new default-implementation path component with its path literal being
+	 * the given {@code source}.
 	 *
-	 * @param source the host sequence to be parsed into a new host.
-	 * @return a host from parsing the given {@code source}.
+	 * @param source the source of the path literal of the constructed path component.
+	 * @return a new path from parsing the given {@code source}.
 	 * @throws NullPointerException     if the given {@code source} is null.
 	 * @throws IllegalArgumentException if the given {@code source} does not match {@link
-	 *                                  UriRegExp#HOST}.
-	 * @since 0.0.1 ~2021.03.20
+	 *                                  UriRegExp#PATH}.
+	 * @since 0.0.1 ~2021.03.21
 	 */
 	@NotNull
 	@Contract(value = "_->new", pure = true)
-	public static Host parse(@NotNull @Pattern(UriRegExp.HOST) String source) {
+	public static Path parse(@NotNull @Pattern(UriRegExp.PATH) String source) {
 		Objects.requireNonNull(source, "source");
-		if (!UriPattern.HOST.matcher(source).matches())
-			throw new IllegalArgumentException("invalid host: " + source);
-		return new Host(source);
+		if (!UriPattern.PATH.matcher(source).matches())
+			throw new IllegalArgumentException("invalid path: " + source);
+		return new Path(source);
 	}
 
 	/**
-	 * Two hosts are equal when they are the same instance or have the same {@code
-	 * host-literal} (the value returned from {@link #toString()}).
+	 * Two paths are equal when they are the same instance or have the same {@code
+	 * path-literal} (the value returned from {@link #toString()}).
 	 *
 	 * @param object the object to be checked.
-	 * @return if the given {@code object} is a host and equals this.
+	 * @return if the given {@code object} is a path and equals this.
 	 * @since 0.0.1 ~2021.03.23
 	 */
 	@Override
@@ -106,19 +102,19 @@ public class Host implements Serializable {
 	public boolean equals(@Nullable Object object) {
 		if (object == this)
 			return true;
-		if (object instanceof Host) {
-			Host host = (Host) object;
+		if (object instanceof Path) {
+			Path path = (Path) object;
 
-			return Objects.equals(this.value, host.toString());
+			return Objects.equals(this.value, path.toString());
 		}
 
 		return false;
 	}
 
 	/**
-	 * The hash code of a host is the hash code of its host-literal. (optional)
+	 * The hash code of a path is the hash code of its path-literal. (optional)
 	 *
-	 * @return the hash code of this host.
+	 * @return the hash code of this path.
 	 * @since 0.0.1 ~2021.03.23
 	 */
 	@Override
@@ -128,20 +124,20 @@ public class Host implements Serializable {
 	}
 
 	/**
-	 * A string representation of this Host. Invoke to get the text representing this in a
+	 * A string representation of the Path. Invoke to get the text representing this in a
 	 * request.
 	 * <br>
 	 * Example:
 	 * <pre>
-	 *     www.example.com
+	 *     /forum/questions/
 	 * </pre>
 	 *
-	 * @return a string representation of this Host.
+	 * @return a string representation of the Path.
 	 * @since 0.0.1 ~2021.03.20
 	 */
 	@NotNull
 	@Contract(pure = true)
-	@Pattern(UriRegExp.HOST)
+	@Pattern(UriRegExp.PATH)
 	@Override
 	public String toString() {
 		return this.value;
