+++ src/main/java/org/cufy/uri/extensions.kt
@@ -13,11 +13,19 @@
  *	See the License for the specific language governing permissions and
  *	limitations under the License.
  */
-/**
- * A package containing standard HTTP related syntax.
- *
- * @author LSafer
- * @version 0.0.1
- * @since 0.0.1 ~2021.03.21
- */
-package org.cufy.http.syntax;
+package org.cufy.uri
+
+/** An alias for [Query.put]. */
+@JvmName("putAt")
+operator fun Query.set(name: String, value: String?): Unit =
+    if (value === null) remove(name) else put(name, value)
+
+/** An alias for [UserInfo.put] and [UserInfo.remove]. */
+@JvmName("putAt")
+operator fun UserInfo.set(index: Int, value: String?): Unit =
+    if (value === null) remove(index) else put(index, value)
+
+/** An alias for [UserInfo.add]. */
+@JvmName("leftShift")
+operator fun UserInfo.plusAssign(value: String): Unit =
+    add(value)
